+++ lib/src/easy_localization_app.dart
@@ -1,4 +1,6 @@
-import 'package:flutter/widgets.dart';
+import 'dart:developer';
+
+import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
@@ -13,42 +15,45 @@ class EasyLocalization extends StatefulWidget {
   final bool useOnlyLangCode;
   final String path;
   final AssetLoader assetLoader;
-
   EasyLocalization({
+    Key key,
     @required this.child,
     @required this.supportedLocales,
-    @required this.fallbackLocale,
     @required this.path,
+    this.fallbackLocale,
     this.useOnlyLangCode = false,
     this.assetLoader = const RootBundleAssetLoader(),
-  })  : assert(supportedLocales.contains(fallbackLocale)),
-        delegate = _EasyLocalizationDelegate(
-          path: path,
-          supportedLocales: supportedLocales,
-          useOnlyLangCode: useOnlyLangCode,
-        );
+  })  : //assert(supportedLocales.contains(fallbackLocale)),
 
+        delegate = _EasyLocalizationDelegate(
+            path: path,
+            supportedLocales: supportedLocales,
+            useOnlyLangCode: useOnlyLangCode,
+            assetLoader: assetLoader),
+        super(key: key);
   _EasyLocalizationState createState() => _EasyLocalizationState();
 
   static _EasyLocalizationState of(BuildContext context) =>
       _EasyLocalizationProvider.of(context).data;
-
-  /// ensureInitialized needs to be called in main
-  /// so that savedLocale is loaded and used from the
-  /// start.
-  static ensureInitialized() async =>
-      await _EasyLocalizationLocale.initSavedAppLocale();
 }
 
 class _EasyLocalizationLocale extends ChangeNotifier {
   Locale _locale;
   static Locale _savedLocale;
+  // Get default OS Locale
+  static final _osCurrentLocale = Intl.getCurrentLocale().split("_");
+  static Locale _osLocal = Locale(_osCurrentLocale[0], _osCurrentLocale[1]);
 
   // @TOGO maybe add assertion to ensure that ensureInitialized has been called and that
   // _savedLocale is set.
-  _EasyLocalizationLocale(Locale fallbackLocale)
-      : this._locale = _savedLocale ?? fallbackLocale;
-
+  _EasyLocalizationLocale(Locale fallbackLocale, List<Locale> supportedLocales)
+      // if fallbackLocale null and default OS Locale in supportedLocales
+      // init by default OS Locale else init by supportedLocales[0]
+      : this._locale = (_savedLocale ?? fallbackLocale) ??
+            supportedLocales.firstWhere((local) => local == _osLocal,
+                orElse: () => supportedLocales.first) {
+    locale = this._locale;
+  }
   Locale get locale => _locale;
   set locale(Locale l) {
     _locale = l;
@@ -70,12 +75,15 @@ class _EasyLocalizationLocale extends ChangeNotifier {
     await _preferences.setString('codeLa', locale.languageCode);
   }
 
-  static initSavedAppLocale() async {
+  static Future<_EasyLocalizationLocale> initSavedAppLocale(
+      Locale fallbackLocale, List<Locale> supportedLocales) async {
     SharedPreferences _preferences = await SharedPreferences.getInstance();
-    final String _codeLang = _preferences.getString('codeLa');
-    final String _codeCoun = _preferences.getString('codeCa');
+    var _codeLang = _preferences.getString('codeLa');
+    var _codeCoun = _preferences.getString('codeCa');
 
     _savedLocale = _codeLang != null ? Locale(_codeLang, _codeCoun) : null;
+    log(_savedLocale.toString(), name: "initSavedAppLocale");
+    return _EasyLocalizationLocale(fallbackLocale, supportedLocales);
   }
 }
 
@@ -86,25 +94,21 @@ class _EasyLocalizationState extends State<EasyLocalization> {
   set locale(Locale l) {
     if (!supportedLocales.contains(l))
       throw new Exception("Locale $l is not supported by this app.");
-
     _locale.locale = l;
   }
 
   List<Locale> get supportedLocales => widget.supportedLocales;
   _EasyLocalizationDelegate get delegate => widget.delegate;
 
+  Locale get fallbackLocale => widget.fallbackLocale;
+
   @override
   void initState() {
-    _locale = _EasyLocalizationLocale(widget.fallbackLocale);
-    _locale.addListener(() {
-      if (mounted) setState(() {});
-    });
     super.initState();
   }
 
   @override
   void dispose() {
-    _locale.dispose();
     super.dispose();
   }
 
@@ -123,10 +127,28 @@ class _EasyLocalizationState extends State<EasyLocalization> {
   }
 
   @override
-  Widget build(BuildContext context) => _EasyLocalizationProvider(
-        data: this,
-        child: widget.child,
-      );
+  Widget build(BuildContext context) {
+    return FutureBuilder<_EasyLocalizationLocale>(
+      future: _EasyLocalizationLocale.initSavedAppLocale(
+          fallbackLocale, supportedLocales),
+      builder: (BuildContext context,
+          AsyncSnapshot<_EasyLocalizationLocale> snapshot) {
+        if (snapshot.hasData) {
+          if (this._locale == null) this._locale = snapshot.data;
+          snapshot.data.addListener(() {
+            if (mounted) setState(() {});
+          });
+          return _EasyLocalizationProvider(
+            data: this,
+            child: widget.child,
+          );
+        } else {
+          // TODO implement [load, error] widget when init locale
+          return Container();
+        }
+      },
+    );
+  }
 }
 
 class _EasyLocalizationProvider extends InheritedWidget {
@@ -150,12 +172,11 @@ class _EasyLocalizationDelegate extends LocalizationsDelegate<Localization> {
   ///  * use only the lang code to generate i18n file path like en.json or ar.json
   final bool useOnlyLangCode;
 
-  _EasyLocalizationDelegate({
-    @required this.path,
-    @required this.supportedLocales,
-    this.useOnlyLangCode = false,
-    this.assetLoader = const RootBundleAssetLoader(),
-  });
+  _EasyLocalizationDelegate(
+      {@required this.path,
+      @required this.supportedLocales,
+      this.useOnlyLangCode = false,
+      this.assetLoader});
 
   @override
   bool isSupported(Locale locale) => supportedLocales.contains(locale);