+++ lib/src/public_ext.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/widgets.dart';
 import 'package:intl/intl.dart';
 
-import 'localization.dart';
+import 'public.dart' as ez;
 
 extension TextTranslateExtension on Text {
-  Text tr({List<String> args, String gender}) =>
-      Text(Localization.instance.tr(this.data, args: args, gender: gender),
+  Text tr({BuildContext context, List<String> args, String gender}) =>
+      Text(ez.tr(this.data, context: context, args: args, gender: gender),
           key: this.key,
           style: this.style,
           strutStyle: this.strutStyle,
@@ -18,8 +18,9 @@ extension TextTranslateExtension on Text {
           maxLines: this.maxLines,
           semanticsLabel: this.semanticsLabel,
           textWidthBasis: this.textWidthBasis);
-  Text plural(dynamic value, {NumberFormat format}) =>
-      Text(Localization.instance.plural(this.data, value, format: format),
+
+  Text plural(dynamic value, {BuildContext context, NumberFormat format}) =>
+      Text(ez.plural(this.data, value, context: context, format: format),
           key: this.key,
           style: this.style,
           strutStyle: this.strutStyle,
@@ -36,7 +37,7 @@ extension TextTranslateExtension on Text {
 
 extension StringTranslateExtension on String {
   tr({List<String> args, String gender}) =>
-      Localization.instance.tr(this, args: args, gender: gender);
+      ez.tr(this, args: args, gender: gender);
   plural(dynamic value, {NumberFormat format}) =>
-      Localization.instance.plural(this, value, format: format);
+      ez.plural(this, value, format: format);
 }