+++ test/utils/test_asset_loaders.dart
@@ -1,6 +1,8 @@
 import 'package:easy_localization/src/asset_loader.dart';
 
 class JsonAssetLoader extends AssetLoader {
+  const JsonAssetLoader();
+
   @override
   Future<Map<String, dynamic>> load(String localePath) {
     return Future.value({